Add file via upload
This business card is a sleek and modern digital representation created using HTML and CSS. It features a minimalist design with clean lines and a professional layout. The card includes sections for the individual's name, job title, location,E-mail, all styled with CSS for a cohesive and visually appealing look. The use of responsive design ensures that the card is optimized for various devices, maintaining its elegant appearance on both desktop and mobile screens. Subtle hover effects and smooth transitions add a touch of interactivity, making this digital business card both functional and engaging.